Checkpoint adding some changes
